pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyside6>=6.11.0",
    "openpyxl>=3.1.0",
    "odfpy>=1.4.1",
    "pypdf>=6.10.2",
    "opencv-python>=4.13.0.92",
    "pymupdf>=1.27.2.3",
]

[project.optional-dependencies]
ocr-ensemble = [
    "pytesseract>=0.3.13",
    "easyocr>=1.7.2",
    "paddleocr>=2.10.0",
    "transformers>=4.57.1",
    "torch>=2.8.0",
    "pillow>=11.3.0",
    "python-doctr>=0.11.0",
]

[project.scripts]
grader = "ai_final_project.main:main"

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["ai_final_project*"]
